platform-tui/requirements.txt
textual==0.47.1
psycopg2-binary==2.9.9
bcrypt==4.1.2
requests==2.31.0